requirements.txt
tensorflow
streamlit
numpy
pillow
pyserial
opencv-python